requirements.txt
numpy
scipy
scikit-learn
Pillow
matplotlib
gradio